add websocket server

+++ database/migrations/2025_11_22_135614_add_equipment_category_id_to_equipment_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('equipment', function (Blueprint $table) {
-            $table->foreignId('equipment_category_id')->nullable()->after('category_id')->constrained('equipment_categories')->onDelete('set null');
+            if (!Schema::hasColumn('equipment', 'equipment_category_id')) {
+                $table->foreignId('equipment_category_id')->nullable()->after('category_id')->constrained('equipment_categories')->onDelete('set null');
+            }
         });
     }
